some fixes

+++ src/shared.h
@@ -11,7 +11,7 @@
 #include "structure.h"
 #include "error.h"
 
-int check_if_table_exists(char *file_path);
+uint8_t check_if_table_exists(char *file_path);
 void get_headers_structure(FILE *fptr, COLUMN_COUNTER column_number, int *reserve_array);
 t_header read_for_structures(FILE *pfile, uint32_t offset);
 void free_headers_struct(COLUMN_COUNTER column_number, t_header **hstr);
